% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_formatting.R
\name{apply_formatting}
\alias{apply_formatting}
\title{Repeatedly gsub over a document}
\usage{
apply_formatting(source, pattern, replacement)
}
\arguments{
\item{source}{A character vector}

\item{pattern}{A vector of patterns}

\item{replacement}{A vector of replacement values}
}
\value{
A modified version of \code{source}.
}
\description{
Introduce formatting by repeatedly gsub-ing over a source text
}
