% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{billSection}
\alias{billSection}
\title{Section 2404 from S.744 (113th Congress).}
\format{Contains a "document" object (length-1 character vector),
and a "test_data" object (a two column data.frame where the first column
records the terms to be highlighted and the second records their class).}
\source{
\url{https://www.congress.gov/bill/113th-congress/senate-bill/744}
}
\description{
One section from the Border Security, Economic Opportunity, and Immigration
Modernization Act.
}
\keyword{datasets}
