% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_console.R, R/format_html.R,
%   R/format_latex.R
\name{format_console}
\alias{format_console}
\alias{format_html}
\alias{format_latex}
\title{Replacement formatters}
\usage{
format_console(color = NULL, background = NULL, bold = NULL,
  italic = NULL, underline = NULL)

format_html(color = NULL, background = NULL, bold = NULL, italic = NULL,
  underline = NULL)

format_latex(color = NULL, background = NULL, bold = NULL,
  italic = NULL, underline = NULL)
}
\arguments{
\item{color}{A vector of text colors}

\item{background}{A vector of background colors}

\item{bold}{A logical vector}

\item{italic}{A logical vector}

\item{underline}{A logical vector}
}
\value{
A length-2 list
}
\description{
Construct a list of replacement formatting for a given markup format
}
\examples{
# html
format_html(color = c("black", "red", "green"))
# latex
format_latex(background = c("black", "red", "green"))

}
