% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{textcolor-package}
\alias{textcolor-package}
\alias{textcolor}
\title{textcolor}
\description{
A package
}
\details{
This is does something
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
