% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortext.R
\name{colorize}
\alias{colorize}
\alias{colorise}
\title{Colorize a source text}
\usage{
colorize(data, text, ...)

colorise(data, text, ...)
}
\description{
Summarize a text model by visualizing a source text
}
